+++ CHANGELOG.md
@@ -4,11 +4,15 @@
 
 ### Fixes
 
-- `[jest-worker]` Make sure to work with Node TS typings v12 ([#10336](https://github.com/facebook/jest/pull/10336))
-
 ### Chore & Maintenance
 
 ### Performance
+
+## 26.2.1
+
+### Fixes
+
+- `[jest-worker]` Make sure to work with Node TS typings v12 ([#10336](https://github.com/facebook/jest/pull/10336))
 
 ## 26.2.0